requirements.txt
streamlit
requests
biopython